+++ package.json
@@ -83,6 +83,7 @@
     "@types/classnames": "^2.2.10",
     "@types/faker": "^5.1.0",
     "@types/jest": "^26.0.0",
+    "@types/jest-axe": "^3.5.1",
     "@types/lodash": "^4.14.155",
     "@types/node": "^14.0.13",
     "@types/react": "^16.9.35",
@@ -107,6 +108,7 @@
     "faker": "^5.1.0",
     "husky": "^4.3.0",
     "jest": "^26.1.0",
+    "jest-axe": "^4.1.0",
     "jest-css-modules-transform": "^4.0.0",
     "jest-date-mock": "^1.0.8",
     "lint-staged": "^10.4.0",
