+++ src/setupTests.ts
@@ -1,7 +1,10 @@
 import '@testing-library/jest-dom';
 import './tests/initI18n';
+import { toHaveNoViolations } from 'jest-axe';
 
 import { publicRuntimeConfig } from '../next.config';
+
+expect.extend(toHaveNoViolations);
 
 jest.mock('next/config', () => {
   return () => ({
